refactor(audio): move MIDI track listing and background parsing to @webgamekit/audio
Adds packages/audio/src/midi.ts with:
- midiGetTracks(buffer): MidiTrackInfo[] — list tracks with instrument name
- midiParseBackground(buffer, volume?): ScheduledNote[] — all notes at real pitch/duration

parseMidi.ts now re-exports these under the original names (getMidiTracks,
parseAllMidiNotes) so no callers change. Game-specific parseMidiTrack
(lane mapping, difficulty filtering) stays in the game.
