+++ packages/audio/package.json
@@ -23,6 +23,9 @@
   "publishConfig": {
     "access": "public"
   },
+  "dependencies": {
+    "@tonejs/midi": "^2.0.28"
+  },
   "devDependencies": {
     "typescript": "^5.0.0",
     "vite": "^5.0.0"
