+++ src/views/Games/RhythmGame/parseMidi.ts
@@ -1,11 +1,13 @@
 import { Midi } from '@tonejs/midi'
-import { midiNoteToFreq } from '@webgamekit/audio'
+import { midiGetTracks } from '@webgamekit/audio'
 import type { ScheduledNote } from '@webgamekit/audio'
 import type { RhythmNote, RgLane, RgDifficulty } from './config'
 
-const MIN_GAP_MS: Record<RgDifficulty, number> = { easy: 450, medium: 180, hard: 0 }
+export type { MidiTrackInfo } from '@webgamekit/audio'
+export { midiGetTracks as getMidiTracks } from '@webgamekit/audio'
+export { midiParseBackground as parseAllMidiNotes } from '@webgamekit/audio'
 
-export type MidiTrackInfo = { index: number; name: string; noteCount: number }
+const MIN_GAP_MS: Record<RgDifficulty, number> = { easy: 450, medium: 180, hard: 0 }
 
 const applyDifficulty = (notes: RhythmNote[], difficulty: RgDifficulty): RhythmNote[] => {
   const minGap = MIN_GAP_MS[difficulty]
@@ -33,24 +35,6 @@ const trackToNotes = (track: Midi['tracks'][number], difficulty: RgDifficulty): 
 }
 
 /**
- * List all tracks in a MIDI file that contain at least one note.
- * @param buffer - Raw MIDI binary data.
- * @returns Track metadata sorted by note count descending.
- */
-const resolveTrackLabel = (track: Midi['tracks'][number], index: number): string => {
-  const parts = [track.name, track.instrument?.name].filter(Boolean)
-  return parts.length > 0 ? parts.join(' · ') : `Track ${index + 1}`
-}
-
-export const getMidiTracks = (buffer: ArrayBuffer): MidiTrackInfo[] => {
-  const midi = new Midi(buffer)
-  return midi.tracks
-    .map((t, i) => ({ index: i, name: resolveTrackLabel(t, i), noteCount: t.notes.length }))
-    .filter((t) => t.noteCount > 0)
-    .sort((a, b) => b.noteCount - a.noteCount)
-}
-
-/**
  * Parse a specific track from a MIDI ArrayBuffer into a RhythmNote array.
  * @param buffer - Raw MIDI binary data.
  * @param trackIndex - Index of the track to parse (from getMidiTracks).
@@ -75,7 +59,7 @@ export const parseMidiTrack = (
  * @returns Parsed and filtered note array sorted by time.
  */
 export const parseMidiBuffer = (buffer: ArrayBuffer, difficulty: RgDifficulty): RhythmNote[] => {
-  const tracks = getMidiTracks(buffer)
+  const tracks = midiGetTracks(buffer)
   if (tracks.length === 0) throw new Error('No notes found in MIDI file')
   return parseMidiTrack(buffer, tracks[0].index, difficulty)
 }
@@ -89,24 +73,5 @@ export const parseMidiBuffer = (buffer: ArrayBuffer, difficulty: RgDifficulty): 
 export const parseMidiFile = async (file: File, difficulty: RgDifficulty): Promise<RhythmNote[]> =>
   parseMidiBuffer(await file.arrayBuffer(), difficulty)
 
-/**
- * Convert every note in every MIDI track into audio-ready ScheduledNotes using
- * real pitches and actual note durations. Intended for background playback.
- * @param buffer - Raw MIDI binary data.
- * @returns All notes across all tracks as ScheduledNote[], sorted by time.
- */
-export const parseAllMidiNotes = (buffer: ArrayBuffer): ScheduledNote[] => {
-  const midi = new Midi(buffer)
-  return midi.tracks
-    .flatMap((track) =>
-      track.notes.map((n) => ({
-        time: Math.round(n.time * 1000),
-        freq: midiNoteToFreq(n.midi),
-        duration: Math.max(50, Math.round(n.duration * 1000)),
-        volume: 0.12,
-        waveType: 'sine' as const,
-        attackTime: 0.01
-      }))
-    )
-    .sort((a, b) => a.time - b.time)
-}
+// Re-export package type for callers that used to import it from here
+export type { ScheduledNote }
